requirements.txt
asgiref==3.2.10
astroid==2.2.5
certifi==2020.6.20
colorama==0.4.1
cycler==0.10.0
django==3.1.2
django-crispy-forms==1.9.2
django-heroku==0.3.1
gunicorn==20.0.4
isort==4.3.21
kiwisolver==1.3.0
lazy-object-proxy==1.4.1
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.5
Pillow==8.0.1
plotly==4.12.0
psycopg2==2.8.5
pygame==1.9.6
pylint==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.12.0
sqlparse==0.4.1
typed-ast==1.4.0
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.11.2